client/requirements.txt
customtkinter
packaging
Pillow
CTkToolTip
